+++ ingress-controller/pkg/manager/instances/registry_test.go
@@ -1,4 +1,4 @@
-package multiinstance_test
+package instances_test
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"go.uber.org/goleak"
 
 	"github.com/kong/kong-operator/v2/ingress-controller/pkg/manager"
-	"github.com/kong/kong-operator/v2/ingress-controller/pkg/manager/multiinstance"
+	"github.com/kong/kong-operator/v2/ingress-controller/pkg/manager/instances"
 )
 
 const (
@@ -18,46 +18,46 @@ const (
 	tickTime = time.Millisecond * 10
 )
 
-func TestManager_Scheduling(t *testing.T) {
+func TestRegistry_Scheduling(t *testing.T) {
 	onCleanupVerifyThereAreNoLeakedGoroutines(t)
 
 	// Such context will be canceled just before the test ends (Cleanups are run)
 	// so we can ensure all goroutines are cleaned up.
 	ctx := t.Context()
 
-	multiManager := multiinstance.NewManager(testr.New(t))
+	registry := instances.NewRegistry(testr.New(t))
 
 	mockInstance1 := newMockInstance(manager.NewRandomID())
 	mockInstance2 := newMockInstance(manager.NewRandomID())
 
-	t.Run("can schedule instances before starting the manager", func(t *testing.T) {
-		err := multiManager.ScheduleInstance(mockInstance1)
+	t.Run("can schedule instances before starting the registry", func(t *testing.T) {
+		err := registry.ScheduleInstance(mockInstance1)
 		require.NoError(t, err)
 
-		err = multiManager.ScheduleInstance(mockInstance2)
+		err = registry.ScheduleInstance(mockInstance2)
 		require.NoError(t, err)
 
-		require.False(t, mockInstance1.wasStarted.Load(), "instance should not have been started yet as the manager is not running")
+		require.False(t, mockInstance1.wasStarted.Load(), "instance should not have been started yet as the registry is not running")
 	})
 
 	t.Run("scheduling an instance with the same ID should fail", func(t *testing.T) {
-		err := multiManager.ScheduleInstance(mockInstance1)
-		require.ErrorIs(t, err, multiinstance.NewInstanceWithIDAlreadyScheduledError(mockInstance1.ID()))
+		err := registry.ScheduleInstance(mockInstance1)
+		require.ErrorIs(t, err, instances.NewInstanceWithIDAlreadyScheduledError(mockInstance1.ID()))
 	})
 
-	managerRunning := make(chan struct{})
-	t.Run("can run the manager", func(t *testing.T) {
+	registryRunning := make(chan struct{})
+	t.Run("can run the registry", func(t *testing.T) {
 		go func() {
-			close(managerRunning)
-			assert.NoError(t, multiManager.Start(ctx))
+			close(registryRunning)
+			assert.NoError(t, registry.Start(ctx))
 		}()
 	})
 
-	t.Run("can schedule instances after starting the manager", func(t *testing.T) {
-		<-managerRunning // Wait for the manager to start.
+	t.Run("can schedule instances after starting the registry", func(t *testing.T) {
+		<-registryRunning // Wait for the registry to start.
 
 		mockInstance3 := newMockInstance(manager.NewRandomID())
-		err := multiManager.ScheduleInstance(mockInstance3)
+		err := registry.ScheduleInstance(mockInstance3)
 		require.NoError(t, err)
 
 		require.EventuallyWithT(t, func(t *assert.CollectT) {
@@ -66,7 +66,7 @@ func TestManager_Scheduling(t *testing.T) {
 	})
 
 	t.Run("can stop an instance", func(t *testing.T) {
-		err := multiManager.StopInstance(mockInstance1.ID())
+		err := registry.StopInstance(mockInstance1.ID())
 		require.NoError(t, err)
 
 		require.EventuallyWithT(t, func(t *assert.CollectT) {
@@ -75,38 +75,38 @@ func TestManager_Scheduling(t *testing.T) {
 	})
 
 	t.Run("can inspect instance readiness", func(t *testing.T) {
-		err := multiManager.IsInstanceReady(mockInstance2.ID())
+		err := registry.IsInstanceReady(mockInstance2.ID())
 		require.NoError(t, err)
 
-		// Deletion from the instance map happens asynchronously in manager
+		// Deletion from the instance map happens asynchronously in the registry
 		// so use require.EventuallyWithT to wait for the error to be returned.
 		// Otherwise it may happen that the error is not returned immediately
 		// because the information hasn't been sent on StopChannel yet.
 		require.EventuallyWithT(t, func(t *assert.CollectT) {
-			err := multiManager.IsInstanceReady(mockInstance1.ID())
-			require.ErrorIs(t, err, multiinstance.NewInstanceNotFoundError(mockInstance1.ID()))
+			err := registry.IsInstanceReady(mockInstance1.ID())
+			require.ErrorIs(t, err, instances.NewInstanceNotFoundError(mockInstance1.ID()))
 		}, waitTime, tickTime)
 	})
 }
 
-func TestManager_WithDiagnosticsExposer(t *testing.T) {
+func TestRegistry_WithDiagnosticsExposer(t *testing.T) {
 	onCleanupVerifyThereAreNoLeakedGoroutines(t)
 
 	ctx := t.Context()
 
-	t.Log("Configuring a manager with a diagnostics exposer")
+	t.Log("Configuring a registry with a diagnostics exposer")
 	diagnosticsExposer := newMockDiagnosticsExposer()
-	multiManager := multiinstance.NewManager(testr.New(t), multiinstance.WithDiagnosticsExposer(diagnosticsExposer))
+	registry := instances.NewRegistry(testr.New(t), instances.WithDiagnosticsExposer(diagnosticsExposer))
 
 	go func() {
-		assert.NoError(t, multiManager.Start(ctx))
+		assert.NoError(t, registry.Start(ctx))
 	}()
 
 	instanceID1 := manager.NewRandomID()
 	instanceID2 := manager.NewRandomID()
 
 	t.Log("Scheduling first instance")
-	err := multiManager.ScheduleInstance(newMockInstance(instanceID1))
+	err := registry.ScheduleInstance(newMockInstance(instanceID1))
 	require.NoError(t, err)
 
 	t.Log("Expecting the diagnostics exposer to have the first instance registered")
@@ -116,7 +116,7 @@ func TestManager_WithDiagnosticsExposer(t *testing.T) {
 	}, waitTime, tickTime)
 
 	t.Log("Scheduling second instance")
-	err = multiManager.ScheduleInstance(newMockInstance(instanceID2))
+	err = registry.ScheduleInstance(newMockInstance(instanceID2))
 	require.NoError(t, err)
 
 	t.Log("Expecting the diagnostics exposer to have both instances registered")
@@ -125,7 +125,7 @@ func TestManager_WithDiagnosticsExposer(t *testing.T) {
 	}, waitTime, tickTime)
 
 	t.Log("Stopping first instance")
-	err = multiManager.StopInstance(instanceID1)
+	err = registry.StopInstance(instanceID1)
 	require.NoError(t, err)
 
 	t.Log("Expecting the diagnostics exposer to have only the second instance registered")
@@ -135,7 +135,7 @@ func TestManager_WithDiagnosticsExposer(t *testing.T) {
 	}, waitTime, tickTime)
 
 	t.Log("Stopping second instance")
-	err = multiManager.StopInstance(instanceID2)
+	err = registry.StopInstance(instanceID2)
 	require.NoError(t, err)
 
 	t.Log("Expecting the diagnostics exposer to have no instances registered")